Struct uri::Uri

pub struct Uri {
    pub scheme: String,
    pub username: Option<String>,
    pub password: Option<String>,
    pub host: Option<String>,
    pub port: Option<u16>,
    pub path: Option<String>,
    pub query: Option<String>,
    pub fragment: Option<String>,
}

Uri represents the data contained in an uri string.

Fields

scheme

Represents the scheme of an uri.

Examples

use uri::Uri;

let uri = Uri::new("https://doc.rust-lang.org/book/README.html").unwrap();
assert_eq!(uri.scheme, "https");

username

Represents the username (authority) of an uri.

Examples

use uri::Uri;

let uri = Uri::new("https://user:pass@doc.rust-lang.org/book/README.html").unwrap();
assert_eq!(uri.username.unwrap(), "user");

password

Represents the password (authority) of an uri.

Examples

use uri::Uri;

let uri = Uri::new("https://user:pass@doc.rust-lang.org/book/README.html").unwrap();
assert_eq!(uri.password.unwrap(), "pass");

host

Represents the host of an uri.

Examples

use uri::Uri;

let uri = Uri::new("https://doc.rust-lang.org/book/README.html").unwrap();
assert_eq!(uri.host.unwrap(), "doc.rust-lang.org");

port

Represents the port of an uri.

Examples

use uri::Uri;

let uri = Uri::new("https://doc.rust-lang.org:1234").unwrap();
assert_eq!(uri.host.unwrap(), "doc.rust-lang.org");
assert_eq!(uri.port, Some(1234));

path

Represents the path of an uri.

Examples

use uri::Uri;

let uri = Uri::new("https:/doc.rust-lang.org/book/README.html").unwrap();
assert_eq!(uri.path.unwrap(), "/book/README.html");

query

Represents the query of an uri.

Examples

use uri::Uri;

let uri = Uri::new("https:/doc.rust-lang.org/?query=value&q=v").unwrap();
assert_eq!(uri.query.unwrap(), "query=value&q=v");

fragment

Represents the fragment of an uri.

Examples

use uri::Uri;

let uri = Uri::new("https:/doc.rust-lang.org/#fragment").unwrap();
assert_eq!(uri.fragment.unwrap(), "fragment");

Methods

impl Uri

fn new(uristr: &str) -> Option<Uri>

Creates Uri object from a given uri string.

Failures

If the string isn't a valid uri the function will return None.

Examples

use uri::Uri;

let uri = match Uri::new("https://doc.rust-lang.org/book/README.html") {
    Some(value) => value,
    None => panic!("Oh no!")
};